cephadm: use a shared smb.conf for clustered smb container sets
Use a shared smb.conf when deploying ctdb enabled containers. There was
a problem updating configs on the ctdb enabled clusters and the issue
was that the configwatch sidecar was not using CTDB, rather it had a
"default" copy of smb.conf that enabled only registry config, but not
CTDB. Examining the cluster this problem was found to be general to all
sidecars that are either sambacc based (not starting smbd, winbindd,
etc) and the smbmetrics sidecar.

Fixes: https://tracker.ceph.com/issues/68322

Resolved Conflicts:
	src/cephadm/cephadmlib/daemons/smb.py

Resolves: rhbz#2315693
Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit a92ca37b73c638fb97ab19a50e6b1328062e10e1)
Signed-off-by: John Mulligan <jmulligan@redhat.com>